refactor: more code deduplication
